+++ app/src/main/java/com/komputerisasi/crud/DataManagement/RmMasuk.kt
@@ -1,11 +1,15 @@
 package com.komputerisasi.crud.DataManagement
 
+import android.app.AlertDialog
 import android.os.Bundle
 import android.widget.Toast
 import androidx.appcompat.app.AppCompatActivity
+import com.komputerisasi.crud.LoginUtama
 import com.komputerisasi.crud.MainActivity
 import com.komputerisasi.crud.R
+import com.komputerisasi.crud.Scanning.ScannRmMasuk
 import com.komputerisasi.crud.adapter.RmMasukAdapter
+import com.komputerisasi.crud.konfigurasi.DatabaseHelper
 import com.komputerisasi.crud.model.*
 import com.komputerisasi.crud.presenter.CrudView
 import com.komputerisasi.crud.presenter.Presenter
@@ -16,6 +20,8 @@ class RmMasuk : AppCompatActivity(), CrudView {
 
     private lateinit var presenter: Presenter
     override fun onCreate(savedInstanceState: Bundle?) {
+        LoginUtama.globalVar = selectDatabase("settingurl")
+
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_rm_masuk)
         presenter = Presenter(this)
@@ -29,7 +35,8 @@ class RmMasuk : AppCompatActivity(), CrudView {
         actionbar.setDisplayHomeAsUpEnabled(true)
 
         btnTambah.setOnClickListener {
-            startActivity<UpdateAddRmMasuk>()
+            startActivity<ScannRmMasuk>()
+            finish()
         }
     }
 
@@ -38,6 +45,17 @@ class RmMasuk : AppCompatActivity(), CrudView {
         startActivity<MainActivity>()
         finish()
         return true
+    }
+
+    private fun selectDatabase(namaseting: String): String {
+        var helper = DatabaseHelper(applicationContext)
+        var db = helper.readableDatabase
+        var dt = db.rawQuery("select * from table_settings where name_setting = '"+namaseting+"'",null)
+        if(dt.moveToNext()){
+            return dt.getString(2)
+        }
+        return ""
+        db.close()
     }
 
     override fun onSuccessGetLogin(data: List<DataLogin>?) {}
@@ -129,13 +147,40 @@ class RmMasuk : AppCompatActivity(), CrudView {
     override fun onSuccessGetRmMasuk(data: List<RmMasukItem>?) {
         rvCategory.adapter = RmMasukAdapter(data,object :RmMasukAdapter.onClickItem{
             override fun clicked(item: RmMasukItem?) {
-                startActivity<UpdateAddRmMasuk>("dataItem" to item)
+                val builder = AlertDialog.Builder(this@RmMasuk)
+                builder.setMessage("Are you sure you want to Delete?")
+                    .setCancelable(false)
+                    .setNegativeButton("No") { dialog, id ->
+                        // Dismiss the dialog
+                        dialog.dismiss()
+                    }
+                    .setPositiveButton("Yes") { dialog, id ->
+                        // Delete selected note from database
+                        startActivity<UpdateAddRmMasuk>("dataItem" to item)
+                        finish()
+                    }
+                val alert = builder.create()
+                alert.show()
+
             }
 
             override fun delete(item: RmMasukItem?) {
-                presenter.hapusDataRmMasuk(item?.IdRmMasuk)
-                startActivity<RmMasuk>()
-                finish()
+                val builder = AlertDialog.Builder(this@RmMasuk)
+                builder.setMessage("Are you sure you want to Delete?")
+                    .setCancelable(false)
+                    .setNegativeButton("No") { dialog, id ->
+                        // Dismiss the dialog
+                        dialog.dismiss()
+                    }
+                    .setPositiveButton("Yes") { dialog, id ->
+                        // Delete selected note from database
+                        presenter.hapusDataRmMasuk(item?.IdRmMasuk)
+                        startActivity<RmMasuk>()
+                        finish()
+                    }
+                val alert = builder.create()
+                alert.show()
+
             }
         })
     }
